Speed up MatchedSequencesEncoder & MatchedReceptorsEncoder through caching. Also removed dataset type mapping for these encoders, as they only make sense for RepertoreDatasets
