+++ test/integration_tests/test_exploratoryAnalysisDesignMatrixExporter.py
@@ -4,7 +4,7 @@ from unittest import TestCase
 
 from immuneML.caching.CacheType import CacheType
 from immuneML.data_model.dataset.RepertoireDataset import RepertoireDataset
-from immuneML.encodings.reference_encoding.MatchedSequencesRepertoireEncoder import MatchedSequencesRepertoireEncoder
+from immuneML.encodings.reference_encoding.MatchedSequencesEncoder import MatchedSequencesEncoder
 from immuneML.environment.Constants import Constants
 from immuneML.environment.EnvironmentSettings import EnvironmentSettings
 from immuneML.environment.LabelConfiguration import LabelConfiguration
@@ -49,7 +49,7 @@ class TestExploratoryAnalysisDesignMatrixExporter(TestCase):
         units = {"named_analysis_4": ExploratoryAnalysisUnit(dataset=dataset,
                                                              report=DesignMatrixExporter(name='report', file_format='csv'),
                                                              label_config=label_config,
-                                                             encoder=MatchedSequencesRepertoireEncoder.build_object(dataset,
+                                                             encoder=MatchedSequencesEncoder.build_object(dataset,
                                                                                                              **{"max_edit_distance": 1,
                                                                                                                 "reference": refs,
                                                                                                                 "reads": "all",
